Move JMXReporter into separate module
