+++ flink-dist/pom.xml
@@ -108,6 +108,13 @@ under the License.
 			<version>${project.version}</version>
 		</dependency>
 
+		<dependency>
+			<groupId>org.apache.flink</groupId>
+			<artifactId>flink-metrics-jmx</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+		
+
 	</dependencies>
 
 	<!-- See main pom.xml for explanation of profiles -->
