+++ flink-metrics/pom.xml
@@ -38,6 +38,7 @@ under the License.
 		<module>flink-metrics-dropwizard</module>
 		<module>flink-metrics-ganglia</module>
 		<module>flink-metrics-graphite</module>
+		<module>flink-metrics-jmx</module>
 		<module>flink-metrics-statsd</module>
 	</modules>
 </project>
